[ADH-5488] Make action finish time nullable (#143)
* [ADH-5488] Make action finish time nullable

* [ADH-5405] Fix rule parser

+++ smart-server/src/test/java/org/smartdata/server/engine/TestCmdletManager.java
@@ -222,7 +222,7 @@ public class TestCmdletManager extends MiniSmartClusterHarness {
     Assert.assertTrue(actionInfo.isFinished());
     Assert.assertTrue(actionInfo.isSuccessful());
     assertEquals(actionInfo.getCreateTime(), startTime);
-    assertEquals(actionInfo.getFinishTime(), finishTime);
+    assertEquals((long) actionInfo.getFinishTime(), finishTime);
     assertEquals(cmdletInfo.getState(), CmdletState.DONE);
 
     cmdletManager.updateStatus(
